returnSkip

+++ Customer.java
@@ -1,21 +1,20 @@
 public class Customer{
     private String name;
-    //private int account_num;
-    public float initial_amount;
+    public double initial_amount;
     
     public String getName() {
         return this.name;
     }
-    
-    public float getInito() {
+  
+
+    public double getInito() {
         return this.initial_amount;
     }
-    
     public void setName(String name) {
         this.name = name;
     }
     
-    public void setInit(float initial_amount) {
+    public void setInit(double initial_amount) {
         this.initial_amount = initial_amount;
     }
